Refactor NMR block into its own app module and restructure blocks into common and base
