+++ pydatalab/pydatalab/apps/__init__.py
@@ -1,2 +1,2 @@
 # This import is required to prevent circular imports for application-specific blocks
-from pydatalab.blocks.blocks import DataBlock  # noqa
+from pydatalab.blocks.base import DataBlock  # noqa
