+++ pydatalab/pydatalab/apps/nmr/__init__.py
@@ -1,0 +1,3 @@
+from .blocks import NMRBlock
+
+__all__ = ("NMRBlock",)
