+++ pydatalab/pydatalab/apps/xrd/blocks.py
@@ -6,7 +6,7 @@ import numpy as np
 import pandas as pd
 from scipy.signal import medfilt
 
-from pydatalab.blocks.blocks import DataBlock
+from pydatalab.blocks.base import DataBlock
 from pydatalab.bokeh_plots import mytheme, selectable_axes_plot
 from pydatalab.file_utils import get_file_info_by_id
 from pydatalab.logger import LOGGER
